Change to suspend & resume workflow
Makes startup time 20 seconds faster

+++ RemoteCreate/launch-box.bat
@@ -1,2 +1,12 @@
-vagrant destroy -f
-vagrant up
+@echo off
+echo Starting...
+vagrant resume | find /i "booted and ready"
+
+IF %ERRORLEVEL% EQU 0 (
+    echo Done, you can start!
+    timeout /t 2
+) else (
+    echo Vagrant resume didn't work, recreating
+    vagrant destroy -f
+    vagrant up
+)
